- разделил логику на несколько файлов - добавил маркеры для карты из папки img - научил компонент передавать данные о городе в map
